Refactor: Rename preparation.py to media_preparation.py
REASON: More descriptive filename that clearly indicates the module's purpose

Changes:
- Renamed createbatch/createbatchlib/preparation.py → media_preparation.py
- Updated import in createbatch/createbatch.py
- Git history preserved using git mv

Module contains:
- prepare_media_file() - copies media to batch structure and updates EXIF
- _find_files_to_copy() - finds alternative format versions (PNG, TIF, RAW)

The name "media_preparation" clearly describes media file preparation for batch processing.
